requirements.txt
langchain_google_genai 
python-dotenv
fastapi
pydantic
langchain
langchain-core
uvicorn
gunicorn